python_env/requirements.txt
uvloop
aioredis
boto3
prometheus_client
PyYAML
docker
psutil
opensearch-py
orjson
requests_aws4auth
GitPython
msgpack-python
aiohttp
fastapi
uvicorn[standard]
python-multipart
jinja2
pytest
pytest-asyncio
